backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
pillow==11.3.0
httpx==0.28.1
